ArtemisComponentManager class

+++ src/Artemis/ArtemisComponentType.h
@@ -28,5 +28,5 @@ struct __ArtemisComponentType {
 
 
 CFStringRef ArtemisComponentTypeGetName(ArtemisComponentTypeRef);
-int ArtemisComponentTypeGetIndex(ArtemisComponentTypeRef);
-int ArtemisComponentTypeGetTaxonomy(ArtemisComponentTypeRef);
+ulong ArtemisComponentTypeGetIndex(ArtemisComponentTypeRef);
+ulong ArtemisComponentTypeGetTaxonomy(ArtemisComponentTypeRef);
